+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## fugit 0.9.7  not yet released
 
+* Wire #deflate into Duration.to_long_s / .to_iso_s / .to_plain_s
+
 
 ## fugit 0.9.6  released 2017-05-24